[Refactor#10] FCM 패키지 구조 정리
