+++ src/main/java/com/swyp/server/infra/fcm/dto/FcmTokenRegisterRequest.java
@@ -1,5 +1,6 @@
-package com.swyp.server.infra.fcm;
+package com.swyp.server.infra.fcm.dto;
 
+import com.swyp.server.infra.fcm.entity.Platform;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
