+++ src/main/java/com/swyp/server/infra/fcm/entity/Platform.java
@@ -1,4 +1,4 @@
-package com.swyp.server.infra.fcm;
+package com.swyp.server.infra.fcm.entity;
 
 public enum Platform {
     ANDROID,
